requirements.txt
absl-py==2.1.0
addict==2.4.0
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
comm @ file:///croot/comm_1709322850197/work
contourpy==1.1.1
craft-text-detector @ git+https://github.com/ria-com/craft-text-detector.git@0734fc81fbe9705cffa6d13c71d2ba240b20b422
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.8
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.8
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.19.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gdown==5.1.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.18
google-auth==2.28.1
google-auth-oauthlib==1.0.0
greenlet==3.0.3
grpcio==1.62.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
imgaug==0.4.0
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources==6.1.2
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1691532092695/work
ipython-genutils==0.2.0
ipywidgets @ file:///croot/ipywidgets_1709574692113/work
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
jmespath==0.10.0
joblib==1.3.2
jsonpickle==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1709322880313/work
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
lmdb==1.4.1
loguru==0.7.2
magickwand==0.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
matplotlib-venn==0.11.10
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv==2.0.1
-e git+https://github.com/open-mmlab/mmdeploy.git@bc75c9d6c8940aa03d0e1e5b5962bd930478ba77#egg=mmdeploy
mmdeploy-runtime-gpu==1.3.1
mmdet==3.1.0
mmengine==0.10.3
-e git+https://github.com/open-mmlab/mmocr.git@b18a09b2f063911a2de70f477aa21da255ff505d#egg=mmocr
model-index==0.1.11
modelhub-client @ git+https://github.com/ria-com/modelhub-client.git@d155a12e1c32a0a65b98c911f1e9e8786d02d93f
multidict==6.0.5
multiprocess==0.70.16
nbformat==5.10.3
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.8.1
onnxruntime-gpu==1.8.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.34
ordered-set==4.1.0
oss2==2.17.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.0.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow @ file:///croot/pillow_1707233021655/work
pkgutil_resolve_name==1.3.10
platformdirs @ file:///croot/platformdirs_1692205439124/work
plotly==5.20.0
prettytable==3.10.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==3.20.2
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.1.1
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==1.8.6
PyTurboJPEG @ git+https://github.com/lilohuang/PyTurboJPEG.git@c9a4973ab48e1e3d421881f76d8b1b2a22669fd2
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
rapidfuzz==3.6.1
referencing==0.34.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.18.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.40.6
shapely==2.0.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
soupsieve==2.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
straug==0.1.2
tabulate==0.9.0
telebot==0.0.5
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2023.7.10
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==1.3.1
torchvision==0.14.1
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.65.2
traitlets==5.14.1
typeguard==4.1.5
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2024.1
ujson==5.9.0
ultralytics @ git+https://github.com/ultralytics/ultralytics.git@16a91a9b6b307b777ec4e8987aa98c2b6e484ebc
urllib3==1.26.18
Wand==0.6.13
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension @ file:///croot/widgetsnbextension_1709322880396/work
yapf==0.40.2
yarl==1.9.4
zipp @ file:///croot/zipp_1704206909481/work
zope.event==5.0
zope.interface==6.2
